refactor: dir arch for bullet-proof-react
